Migrate the checksum package to Typescript (#1583)
