+++ package.json
@@ -10,7 +10,7 @@
     "docs-build": "npm run docs-install && npm run docs-build-tasks && (cd website && npm run build)",
     "docs-serve": "npm run docs-build && (cd website && npm run start)",
     "eslint": "eslint --ext .js --ext .ts .",
-    "remove-locks": "lerna exec 'rm package-lock.json'",
+    "remove-locks": "lerna exec 'rm -f package-lock.json'",
     "install-locks": "lerna exec --no-sort -- npm i --package-lock-only --no-audit",
     "lint": "npm run eslint && lerna run python-lint && npm run lint-md",
     "lint-md": "markdownlint docs/**/*.md docs/*.md tf-modules/**/*.md tf-modules/*.md",
