--- packages/checksum/.npmignore
@@ -1,1 +1,0 @@
-/tests/
