--- packages/checksum/index.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const { generateChecksumFromStream } = require('./checksum');
-const { validateChecksumFromStream } = require('./validate');
-
-module.exports = {
-  generateChecksumFromStream,
-  validateChecksumFromStream
-};
