+++ packages/checksum/package.json
@@ -5,10 +5,19 @@
   "engines": {
     "node": ">=10.16.3"
   },
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "./dist/**/*.js",
+    "./dist/**/*.d.ts"
+  ],
   "scripts": {
+    "build-docs": "npm run tsc && jsdoc2md -t templates/README.md.hbs ./dist/*.js > README.md",
     "test": "ava",
     "test-coverage": "nyc ava",
-    "debug": "NODE_ENV=test node --inspect-brk node_modules/ava/profile.js --serial tests/*.js"
+    "debug": "NODE_ENV=test node --inspect-brk node_modules/ava/profile.js --serial tests/*.js",
+    "tsc": "tsc",
+    "prepare": "npm run tsc"
   },
   "nyc": {
     "exclude": [
@@ -21,6 +30,7 @@
   "publishConfig": {
     "access": "public"
   },
+  "homepage": "https://github.com/nasa/cumulus/tree/master/packages/checksum/README.md",
   "repository": {
     "type": "git",
     "url": "https://github.com/nasa/cumulus"
@@ -32,6 +42,8 @@
   },
   "devDependencies": {
     "ava": "^2.1.0",
-    "nyc": "^14.0.0"
+    "jsdoc-to-markdown": "^5.0.3",
+    "nyc": "^14.0.0",
+    "typescript": "^3.8.3"
   }
 }
