+++ packages/checksum/src/cksum.d.ts
@@ -1,0 +1,4 @@
+declare module 'cksum' {
+  import { Writable } from 'stream';
+  export function stream(cb: (result: Buffer, length: number) => void): Writable;
+}
