+++ packages/checksum/src/index.ts
@@ -1,0 +1,6 @@
+/**
+ * @module checksum
+ */
+
+export { generateChecksumFromStream } from './checksum';
+export { validateChecksumFromStream } from './validate';
