+++ packages/checksum/templates/README.md.hbs
@@ -1,0 +1,33 @@
+# @cumulus/checksum
+
+## Checksum
+
+The `@cumulus/checksum` library provides checksum functionality used by Cumulus
+packages and tasks. Currently the supported input includes file streams, and
+supported checksum algorithms include `cksum` and the algorithms available to
+the `crypto` package, as documented
+[here](https://nodejs.org/api/crypto.html#crypto_crypto_createhash_algorithm_options).
+
+## Usage
+
+```js
+const fs = require('fs');
+const { generateChecksumFromStream } = require('@cumulus/checksum');
+
+const stream = fs.createReadStream('myDataFile.hdf');
+const myCksum = generateChecksumFromStream('cksum', stream);
+```
+
+## API
+
+{{>main}}
+
+## About Cumulus
+
+Cumulus is a cloud-based data ingest, archive, distribution and management
+prototype for NASA's future Earth science data streams.
+
+[Cumulus Documentation](https://nasa.github.io/cumulus)
+
+---
+Generated automatically using `npm run build-docs`
